infer source profile to avoid having to set profile envar
Profiles can have a `source_profile` key which indicates where the credentials should come from. Before this commit, this key was ignored, and the default profile was used instead. When the default profile is not valid, you would have to do something like `export AWS_PROFILE=my_source_profile` for the sts call to work; otherwise it would fail with a permission denial.

We now attempt to use the source profile given in the config. If there's none, the `source_profile` Python variable will be empty and I believe the boto session will use the default profile as a result. Not sure if this actually occurs often in normal use.