fix logging + init log at module init
The log was only in scope in the `cli` method which obviously isn't helpful. I meant for the log setup to only be active when the program is ran as a CLI, not when the module is merely imported, but I'm not sure how to do that. For now I'll just make the logging config mandatory so that I have at least some way of emitting logs.