QOL overhaul (#13)
* add missing deps

* remove unnecessary version string

It's already defined in the toml file.

* infer source profile to avoid having to set profile envar

Profiles can have a `source_profile` key which indicates where the credentials should come from. Before this commit, this key was ignored, and the default profile was used instead. When the default profile is not valid, you would have to do something like `export AWS_PROFILE=my_source_profile` for the sts call to work; otherwise it would fail with a permission denial.

We now attempt to use the source profile given in the config. If there's none, the `source_profile` Python variable will be empty and I believe the boto session will use the default profile as a result. Not sure if this actually occurs often in normal use.

* rename objects for clarity

* fix logging + init log at module init

The log was only in scope in the `cli` method which obviously isn't helpful. I meant for the log setup to only be active when the program is ran as a CLI, not when the module is merely imported, but I'm not sure how to do that. For now I'll just make the logging config mandatory so that I have at least some way of emitting logs.

* improve logging + log expiration time

* trivial reformat

* log the request as well

* support custom session name

* add support for duration + duration caching

* add debug log toggle

* bump version

* update install command

* remove placeholder tests

With the version removed, the tests are now technically broken. I haven't had time yet to develop proper tests so I will delete these until I do.

Co-authored-by: Azat Akhmetov <aakhmetov@whipmedia.com>